Corrigindo validação da senha

+++ src/views/SingIn.vue
@@ -1,16 +1,16 @@
 <template>
     <v-container class="d-flex align-center bg-grey-lighten-1 justify-center h-100">
         <v-card elevation="10" width="320" height="320" class="d-flex align-center">
-        <v-sheet class="mx-auto" width="300">
-            <v-form ref="form" fast-fail @submit.prevent>
-                <v-text-field v-model="email" :rules="emailRules" label="E-mail"></v-text-field>
-                <v-text-field v-model="password" :rules="passwordRules" label="Senha"></v-text-field>
+            <v-sheet class="mx-auto" width="300">
+                <v-form ref="form" fast-fail @submit.prevent>
+                    <v-text-field v-model="email" :rules="emailRules" label="E-mail"></v-text-field>
+                    <v-text-field v-model="password" :rules="passwordRules" label="Senha"></v-text-field>
 
-                <v-btn class="mt-2" color="success" type="submit" block @click="validate">Logar</v-btn>
-                <v-btn class="mt-2" color="primary" type="submit" block @click="signIn">Cadastrar</v-btn>
-            </v-form>
-        </v-sheet>
-    </v-card>
+                    <v-btn class="mt-2" color="success" type="submit" block @click="validate">Logar</v-btn>
+                    <v-btn class="mt-2" color="primary" type="submit" block @click="registerUser">Cadastrar</v-btn>
+                </v-form>
+            </v-sheet>
+        </v-card>
     </v-container>
 </template>
 
@@ -46,20 +46,28 @@ export default {
         validate() {
             if (this.$refs.form.validate()) {
                 // Validação do e-mail
-                const emailExists = this.checkEmailExists(this.email);
-                const passwordExists = this.checkPasswordExists(this.password);
-                if (emailExists && passwordExists) {
-                    // Autenticado com sucesso, redirecionar para a página de login
-                   
-                    this.$router.push("/");
-                    // Disparar o evento isAuthenticated para true
-                    EventBus.isAuthenticated = true;
-                    console.log("Valor de isAuthenticated:", EventBus.isAuthenticated);
-                   
-
+                const emailValid = this.emailRules.every(rule => rule(this.email) === true);
+                if (emailValid) {
+                    const emailExists = this.checkEmailExists(this.email);
+                    if (emailExists) {
+                        const passwordValid = this.passwordRules.every(rule => rule(this.password) === true);
+                        if (passwordValid) {
+                            // Autenticado com sucesso, redirecionar para a página de login
+                            this.$router.push("/");
+                            // Disparar o evento isAuthenticated para true
+                            EventBus.isAuthenticated = true;
+                            console.log("Valor de isAuthenticated:", EventBus.isAuthenticated);
+                        } else {
+                            // Senha inválida
+                            alert('Senha inválida. Por favor, verifique as regras de senha.');
+                        }
+                    } else {
+                        // E-mail não cadastrado
+                        alert('Usuário não cadastrado, por favor cadastre-se.');
+                    }
                 } else {
-                    // E-mail não cadastrado
-                    alert('Usuario não cadastro, por favor cadastrar');
+                    // E-mail inválido
+                    alert('E-mail inválido. Por favor, verifique o e-mail inserido.');
                 }
             }
         },
@@ -68,23 +76,32 @@ export default {
             const users = JSON.parse(localStorage.getItem('users')) || [];
             return users.some(user => user.email === email);
         },
-        checkPasswordExists(password) {
-            // Verifica se a senha já está armazenada no LocalStorage
-            const users = JSON.parse(localStorage.getItem('users')) || [];
-            return users.some(user => user.password === password);
-        },
         registerUser() {
-            // Registra um novo usuário no LocalStorage
-            const newUser = { email: this.email, password: this.password };
             const users = JSON.parse(localStorage.getItem('users')) || [];
-            users.push(newUser);
-            localStorage.setItem('users', JSON.stringify(users));
-            alert('Usuário cadastrado com sucesso!');
+            if (this.email && this.password) {
+                const emailValid = this.emailRules.every(rule => rule(this.email) === true);
+                if (emailValid) {
+                    const passwordValid = this.passwordRules.every(rule => rule(this.password) === true);
+                    if (passwordValid) {
+                        // Verifica se o e-mail já existe antes de registrar
+                        const emailExists = this.checkEmailExists(this.email);
+                        if (!emailExists) {
+                            users.push({ email: this.email, password: this.password });
+                            localStorage.setItem('users', JSON.stringify(users));
+                            alert('Usuário cadastrado com sucesso!');
+                        } else {
+                            alert('Este e-mail já está cadastrado.');
+                        }
+                    } else {
+                        alert('Senha inválida. Por favor, verifique as regras de senha.');
+                    }
+                } else {
+                    alert('E-mail inválido. Por favor, verifique o e-mail inserido.');
+                }
+            } else {
+                alert('Os campos de e-mail e senha devem ser preenchidos corretamente.');
+            }
         },
-        signIn() {
-            // Se o e-mail não existir no LocalStorage, pode prosseguir com o cadastro
-            this.registerUser();
-        }
     }
 }
 </script>
